Fix LineMOD compilation errors
