+++ recognition/include/pcl/recognition/color_gradient_dot_modality.h
@@ -42,76 +42,12 @@
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
 
+#include <pcl/recognition/point_types.h>
 #include <pcl/recognition/dot_modality.h>
 #include <pcl/recognition/quantized_map.h>
 
-
 namespace pcl
 {
-
-  /** \brief A point structure for representing RGB color
-    * \ingroup common
-    */
-  struct EIGEN_ALIGN16 PointRGB
-  {
-    union
-    {
-      union
-      {
-        struct
-        {
-          uint8_t b;
-          uint8_t g;
-          uint8_t r;
-          uint8_t _unused;
-        };
-        float rgb;
-      };
-      uint32_t rgba;
-    };
-
-    inline PointRGB ()
-    {}
-
-    inline PointRGB (const uint8_t b, const uint8_t g, const uint8_t r)
-      : b (b), g (g), r (r), _unused (0)
-    {}
-
-    EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-  };
-
-
-  /** \brief A point structure representing Euclidean xyz coordinates, and the intensity value.
-    * \ingroup common
-    */
-  struct EIGEN_ALIGN16 GradientXY
-  {
-    union
-    {
-      struct
-      {
-        float x;
-        float y;
-        float angle;
-        float magnitude;
-      };
-      float data[4];
-    };
-    EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-
-    inline bool operator< (const GradientXY & rhs)
-    {
-      return (magnitude > rhs.magnitude);
-    }
-  };
-  inline std::ostream & operator << (std::ostream & os, const GradientXY & p)
-  {
-    os << "(" << p.x << "," << p.y << " - " << p.magnitude << ")";
-    return (os);
-  }
-
-  // --------------------------------------------------------------------------
-
   template <typename PointInT>
   class ColorGradientDOTModality
     : public DOTModality, public PCLBase<PointInT>
