+++ recognition/include/pcl/recognition/point_types.h
@@ -49,33 +49,33 @@ namespace pcl
   /** \brief A point structure for representing RGB color
     * \ingroup common
     */
-  //struct EIGEN_ALIGN16 PointRGB
-  //{
-  //  union
-  //  {
-  //    union
-  //    {
-  //      struct
-  //      {
-  //        uint8_t b;
-  //        uint8_t g;
-  //        uint8_t r;
-  //        uint8_t _unused;
-  //      };
-  //      float rgb;
-  //    };
-  //    uint32_t rgba;
-  //  };
+  struct EIGEN_ALIGN16 PointRGB
+  {
+   union
+   {
+     union
+     {
+       struct
+       {
+         uint8_t b;
+         uint8_t g;
+         uint8_t r;
+         uint8_t _unused;
+       };
+       float rgb;
+     };
+     uint32_t rgba;
+   };
 
-  //  inline PointRGB ()
-  //  {}
+   inline PointRGB ()
+   {}
 
-  //  inline PointRGB (const uint8_t b, const uint8_t g, const uint8_t r)
-  //    : b (b), g (g), r (r), _unused (0)
-  //  {}
+   inline PointRGB (const uint8_t b, const uint8_t g, const uint8_t r)
+     : b (b), g (g), r (r), _unused (0)
+   {}
 
-  //  EIGEN_MAKE_ALIGNED_OPERATOR_NEW
-  //};
+   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
+  };
 
 
   /** \brief A point structure representing Euclidean xyz coordinates, and the intensity value.
